[mdZ80] ReadB/WriteB functions should use uint16_t for address.

+++ src/libgens/cpu/Z80_MD_Mem.cpp
@@ -78,7 +78,7 @@ void Z80_MD_Mem::End(void)
  * @param address Address to read from.
  * @return YM2612 register.
  */
-inline uint8_t Z80_MD_Mem::Z80_ReadB_YM2612(uint32_t address)
+inline uint8_t Z80_MD_Mem::Z80_ReadB_YM2612(uint16_t address)
 {
 	// According to the Genesis Software Manual, all four addresses return
 	// the same value for YM2612_Read().
@@ -99,7 +99,7 @@ inline uint8_t Z80_MD_Mem::Z80_ReadB_YM2612(uint32_t address)
  * @param address Address to read from.
  * @return VDP register.
  */
-inline uint8_t Z80_MD_Mem::Z80_ReadB_VDP(uint32_t address)
+inline uint8_t Z80_MD_Mem::Z80_ReadB_VDP(uint16_t address)
 {
 	if (address < 0x7F00 || address > 0x7F1F)
 	{
@@ -158,7 +158,7 @@ inline uint8_t Z80_MD_Mem::Z80_ReadB_VDP(uint32_t address)
  * @param address Address to read from.
  * @return Byte from MC68000 ROM.
  */
-inline uint8_t Z80_MD_Mem::Z80_ReadB_68K_Rom(uint32_t address)
+inline uint8_t Z80_MD_Mem::Z80_ReadB_68K_Rom(uint16_t address)
 {
 	// Z80 cannot read from M68K RAM.
 	// If this is attempted, 0xFF will be returned.
@@ -180,7 +180,7 @@ inline uint8_t Z80_MD_Mem::Z80_ReadB_68K_Rom(uint32_t address)
  * @param address Address to write to.
  * @param data Byte to write.
  */
-inline void Z80_MD_Mem::Z80_WriteB_Bank(uint32_t address, uint8_t data)
+inline void Z80_MD_Mem::Z80_WriteB_Bank(uint16_t address, uint8_t data)
 {
 	if (address > 0x60FF)
 	{
@@ -199,7 +199,7 @@ inline void Z80_MD_Mem::Z80_WriteB_Bank(uint32_t address, uint8_t data)
  * @param address Address to write to.
  * @param data Byte to write.
  */
-inline void Z80_MD_Mem::Z80_WriteB_YM2612(uint32_t address, uint8_t data)
+inline void Z80_MD_Mem::Z80_WriteB_YM2612(uint16_t address, uint8_t data)
 {
 	// The YM2612's RESET line is tied to the Z80's RESET line.
 	if (M68K_Mem::Z80_State & Z80_STATE_RESET)
@@ -215,7 +215,7 @@ inline void Z80_MD_Mem::Z80_WriteB_YM2612(uint32_t address, uint8_t data)
  * @param address Address to write to.
  * @param data Byte to write.
  */
-inline void Z80_MD_Mem::Z80_WriteB_VDP(uint32_t address, uint8_t data)
+inline void Z80_MD_Mem::Z80_WriteB_VDP(uint16_t address, uint8_t data)
 {
 	if (address < 0x7F00 || address > 0x7F1F)
 	{
@@ -254,7 +254,7 @@ inline void Z80_MD_Mem::Z80_WriteB_VDP(uint32_t address, uint8_t data)
  * @param address Address to write to.
  * @param data Byte to write.
  */
-inline void Z80_MD_Mem::Z80_WriteB_68K_Rom(uint32_t address, uint8_t data)
+inline void Z80_MD_Mem::Z80_WriteB_68K_Rom(uint16_t address, uint8_t data)
 {
 	// NOTE: Z80 writes to M68K RAM are allowed.
 	// Reference: http://gendev.spritesmind.net/forum/viewtopic.php?t=985
@@ -273,7 +273,7 @@ inline void Z80_MD_Mem::Z80_WriteB_68K_Rom(uint32_t address, uint8_t data)
  * @param address Address to read from.
  * @return Byte from the Z80 address space.
  */
-uint8_t Z80_MD_Mem::Z80_ReadB(uint32_t address)
+uint8_t Z80_MD_Mem::Z80_ReadB(uint16_t address)
 {
 	const uint8_t page = ((address >> 12) & 0x0F);
 	switch (page & 0x0F)
@@ -313,7 +313,7 @@ uint8_t Z80_MD_Mem::Z80_ReadB(uint32_t address)
  * @param address Address to write to.
  * @param data Byte to write to the Z80 address space.
  */
-void Z80_MD_Mem::Z80_WriteB(uint32_t address, uint8_t data)
+void Z80_MD_Mem::Z80_WriteB(uint16_t address, uint8_t data)
 {
 	const uint8_t page = ((address >> 12) & 0x0F);
 	switch (page & 0x0F)
